requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-gis>=1.1,<2.0
djangorestframework-simplejwt>=5.3,<6.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=23.0,<24.0
afromessage>=0.1.4,<1.0
twilio>=9.0,<10.0
africastalking>=1.2,<2.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==26.1
psycopg2==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyOpenSSL==26.0.0
redis==7.4.0
requests==2.33.1
service-identity==24.2.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
urllib3==2.6.3
websockets==16.0
zope.interface==8.3
